requirements.txt
Django==1.10.5
django-registration
django-background-tasks
